Make Enhanced Monitoring IAM role prefix to be configurable variable
There is a limit of 32 characters for IAM role prefix and terraform will throw an error like this when that happens:

```
Error: module.aurora_hint.aws_iam_role.rds-enhanced-monitoring: "name_prefix" cannot be longer than 32 characters, name is limited to 64
```

The hardcoded prefix was already using up 19 out of 32 allowed characters.

Making this as a configurable variable with backwards compatibility.